+++ lib/open-build-service-api.rb
@@ -1,5 +1,6 @@
 require 'cgi'
 require 'date'
+require 'forwardable'
 require 'net/http'
 require 'nokogiri'
 require_relative 'version'
@@ -9,6 +10,10 @@ require_relative 'connection'
 # OBS API
 require_relative 'api/about'
 require_relative 'api/projects'
+
+# Collections
+require_relative 'collections/abstract_collection'
+require_relative 'collections/packages_collection'
 
 # Models
 require_relative 'models/package'